requirements.txt
alembic==1.7.4
autopep8==1.5.7
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
mysqlclient==2.0.3
protobuf==3.18.1
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.19.1
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==2.0.2
WTForms==2.3.3
